app/requirements.txt
python-telegram-bot
python-dotenv
regex
pymongo[srv]
dnspython